refactor(#741): refactor massifController
